refactor: extract benchmark comparison policy
